requirements.txt
langchain
openai
jupyter
huggingface_hub
streamlit
python-dotenv